+++ pom.xml
@@ -61,7 +61,7 @@
         <module>providers/arangodb/shedlock-provider-arangodb</module>
         <module>providers/ignite/shedlock-provider-ignite</module>
         <module>providers/inmemory/shedlock-provider-inmemory</module>
-        <module>providers/memcached/shedlock-provider-memcached</module>
+        <module>providers/memcached/shedlock-provider-spymemcached</module>
     </modules>
 
     <properties>
