--- providers/memcached/shedlock-provider-memcached/docker/docker-compose.yml
@@ -1,31 +1,0 @@
----
-version: "3.7"
-
-networks:
-  memcached:
-    name: memcached
-  dashboard:
-    name: dashboard
-
-services:
-  memcached:
-    container_name: memcached
-    image: memcached
-    ports:
-      - 11211:11211
-    networks:
-      - memcached
-
-  dashboard:
-    container_name: dashboard
-    image: hatamiarash7/memcached-admin
-    ports:
-      - 8080:80
-    environment:
-      - MEMCACHED_HOST=memcached
-      - MEMCACHED_PORT=11211
-    networks:
-      - memcached
-      - dashboard
-
-...
