+++ providers/memcached/shedlock-provider-spymemcached/pom.xml
@@ -8,7 +8,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>shedlock-provider-memcached</artifactId>
+    <artifactId>shedlock-provider-spymemcached</artifactId>
     <version>4.32.1-SNAPSHOT</version>
 
     <properties>
